+++ commons/src/main/java/commons/SinglePlayer.java
@@ -4,13 +4,11 @@ import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 
-import javax.persistence.*;
-
 import static org.apache.commons.lang3.builder.ToStringStyle.MULTI_LINE_STYLE;
 
-public class SinglePlayer extends Player{
+public class SinglePlayer extends Player {
 
-    public SinglePlayer(String username, int score) {
+    public SinglePlayer(String username, long score) {
         super(username, score);
     }
